+++ CHANGELOG.md
@@ -7,6 +7,7 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 ## Unreleased
 
 - Give nice error if you declare custom scalar with same name as built-in.
+- Re-export juniper from juniper-from-schema to make sure we're always using the same version.
 
 #### Breaking changes